requirements.txt
asgiref==3.2.10
dj-pagination==2.5.0
Django==3.1.2
django-extra-views==0.13.0
django-filter==2.4.0
psycopg2==2.8.6
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
